pyproject.toml
[tool.poetry]
name = "defender2yara"
version = "1.0"
description = "Convert Microsoft Defender Antivirus Signatures(VDM) to YARA rules."
authors = ["Tomoaki Tani"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyexiftool = "^0.5.6"
yara-python = "^4.5.1"
httpx = "^0.27.0"
tqdm = "^4.66.4"
libarchive-c = "^5.1"
toml = "^0.10.2"
colorama = "^0.4.6"
pefile = "^2023.2.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
defender2yara = "defender2yara.__main__:run"